FINAL //com comentários

+++ readSVG.cpp
@@ -19,35 +19,34 @@ namespace svg {
             origin_stream >> x >> y;
             origin = {x, y};
         }
-
         stringstream ss(transform_str);
         string transform;
 
         while (ss >> transform) {
             if (transform.find("translate") == 0) {
-                float tx, ty = 0;
+                int tx, ty = 0;
 
-                // Replace spaces with commas
+                //Replace spaces with commas - not working
                 size_t open_bracket_pos = transform.find("(");
                 size_t comma_pos = transform.find(" ");
                 if (open_bracket_pos != string::npos && comma_pos != string::npos) {
                     transform[open_bracket_pos + 1] = ',';
                 }
-                sscanf(transform.c_str(), "translate(%f %f)", &tx, &ty);
-                Point translation = {static_cast<int>(tx), static_cast<int>(ty)};
+                sscanf(transform.c_str(), "translate(%i %i)", &tx, &ty);
+                Point translation = {tx, ty};
                 element->translate(translation);
             }
             else if (transform.find("rotate") == 0) {
                 float angle;
                 sscanf(transform.c_str(), "rotate(%f)", &angle);
-                element->rotate(origin, static_cast<int>(angle));
+                element->rotate(origin, (angle));
             }
             else if (transform.find("scale") == 0) {
                 float sx, sy = 1;
                 sscanf(transform.c_str(), "scale(%f,%f)", &sx, &sy);
                 if (sx < 0) sx = 0;
                 if (sy < 0) sy = 0;
-                element->scale(origin, static_cast<int>(sx));
+                element->scale(origin, (sx));
             }
         }
     }
@@ -64,26 +63,30 @@ namespace svg {
         dimensions.y = xml_elem->IntAttribute("height");
 
         XMLElement* child = xml_elem->FirstChildElement();
+
+        //Check the type of SVG element & create an instance
         while (child != nullptr) {
             SVGElement* element = nullptr;
             const char* transform_attr = nullptr;
 
             if (strcmp(child->Name(), "circle") == 0) {
-                float cx = child->FloatAttribute("cx");
-                float cy = child->FloatAttribute("cy");
-                float r = child->FloatAttribute("r");
+                //Read attributes for circle element
+                int cx = child->FloatAttribute("cx");
+                int cy = child->FloatAttribute("cy");
+                int r = child->FloatAttribute("r");
                 const char* fill_color = child->Attribute("fill");
 
-                element = new Circle(parse_color(fill_color), {static_cast<int>(cx), static_cast<int>(cy)}, static_cast<int>(r));
+                //Create a circle element instance, similar to other SVG Element types
+                element = new Circle(parse_color(fill_color), {(cx), cy}, r);
             }
             else if (strcmp(child->Name(), "ellipse") == 0) {
-                float cx = child->FloatAttribute("cx");
-                float cy = child->FloatAttribute("cy");
-                float rx = child->FloatAttribute("rx");
-                float ry = child->FloatAttribute("ry");
+                int cx = child->FloatAttribute("cx");
+                int cy = child->FloatAttribute("cy");
+                int rx = child->FloatAttribute("rx");
+                int ry = child->FloatAttribute("ry");
                 const char* fill_color = child->Attribute("fill");
 
-                element = new Ellipse(parse_color(fill_color), {static_cast<int>(cx), static_cast<int>(cy)}, {static_cast<int>(rx), static_cast<int>(ry)});
+                element = new Ellipse(parse_color(fill_color), {cx, cy}, {rx, ry});
             }
             else if (strcmp(child->Name(), "polyline") == 0) {
                 const char* points_str = child->Attribute("points");
@@ -103,13 +106,13 @@ namespace svg {
                 element = new Polyline(parse_color(stroke_color), points);
             }
             else if (strcmp(child->Name(), "line") == 0) {
-                float x1 = child->FloatAttribute("x1");
-                float y1 = child->FloatAttribute("y1");
-                float x2 = child->FloatAttribute("x2");
-                float y2 = child->FloatAttribute("y2");
+                int x1 = child->FloatAttribute("x1");
+                int y1 = child->FloatAttribute("y1");
+                int x2 = child->FloatAttribute("x2");
+                int y2 = child->FloatAttribute("y2");
                 const char* stroke_color = child->Attribute("stroke");
 
-                element = new Line(parse_color(stroke_color), {static_cast<int>(x1), static_cast<int>(y1)}, {static_cast<int>(x2), static_cast<int>(y2)});
+                element = new Line(parse_color(stroke_color), {x1, y1}, {x2, y2});
             }
             else if (strcmp(child->Name(), "polygon") == 0) {
                 const char* points_str = child->Attribute("points");
@@ -139,14 +142,18 @@ namespace svg {
             }
 
             if (element != nullptr) {
+                //Extract transformation attributes
                 transform_attr = child->Attribute("transform");
+
+                //Apply transformations if available
                 const char* transform_origin_attr = child->Attribute("transform-origin");
                 if (transform_attr != nullptr) {
                     applyTransformation(element, transform_attr, transform_origin_attr);
                 }
+                //Add the element to the vector of SVG elements
                 svg_elements.push_back(element);
             }
-
+            //Move to the next child element
             child = child->NextSiblingElement();
         }
     }
